requirements.txt
google-generativeai>=0.8.5
openai>=1.99.1
requests>=2.32.4
